Fix some inconsistencies in python bindings between clang and core type parsers

+++ python/generator.cpp
@@ -340,7 +340,7 @@ int main(int argc, char* argv[])
 			fprintf(enums, "\n\nclass %s(enum.IntEnum):\n", name.c_str());
 			for (auto& j : i.second->GetEnumeration()->GetMembers())
 			{
-				fprintf(enums, "\t%s = %" PRId64 "\n", j.name.c_str(), j.value);
+				fprintf(enums, "\t%s = %" PRId32 "\n", j.name.c_str(), (int32_t)j.value);
 			}
 		}
 		else if ((i.second->GetClass() == BoolTypeClass) || (i.second->GetClass() == IntegerTypeClass)
